Use different iter counts for expand/final optimize
If we get the expansion wrong, nothing is lost. We may expand a better
candidate a bit later, but if they were close, we will look at it soon
anyway. It affects the search order a bit, but it should not affect the
leaves that we reach too much, unless the leaves are very similar in
performance, but in that case it does not matter which one we choose.

For exploring the variations of the final plan though, we may want to
spend some more iterations to be:

 1. Absolutely sure that this is the best statement order. We can spend
    more effort ranking leaves among each other than ranking interior
    nodes.

 2. Sure that we get the most efficient plan for the given statement
    order.

Also in this commit: use more descriptive names, swap order of the plans
because "plans[0]" is shorter than "plans[plans.len() - 1]".