Reorganize crates into workspaces, add ffi crate
I want to start exposign a C API so that Noblit can be used from other
languages than Rust. It woulb be nice to do that on top of the crate,
rather than inside the crate. So move the crate into the "noblit"
directory. That will contain the implementation, in Rust. Then there is
also the "libnoblit" crate, which exposes a C API on top of the Rust
crate. This one produces libnoblit.so and libnoblit.a for use in other
applications. Tie things together with a Cargo workspace file.