Add initial version of Python client library
This just wraps the two functions that I have now, for loading and
freeing. It works, because it does not crash if you pass a valid
database as file name. If you pass an invalid file name, it aborts.

But apart from this, there is no way yet to run queries, or anything.