Add libgmp to Nix dependences
Now that I am taking binutils from Nix, suddenly the "stack setup" step
fails in ld.gold from binutils from the Nix store. Perhaps adding the
binutils package caused a new linker to be on the path, and that linker
does not search outside of he Nix store? In any case, libgmp *is* a
dependency of the Haskell runtime, so it being absent from the devenv
was a real bug.