Do not try to link precompiled header
With ld from binutils 2.31.1, we get an error when trying to link a
precompiled header (which is perhaps appropriate). With binutils 2.33.1,
this error does not occur. Before I added binutils to the devenv, I was
using my local one, which is 2.33.1, but current Nixpkgs ships 2.31.1
apparently.

In any case, the solution is to pass "-c" to GCC so it only compiles, it
does not link. The short form is not great for use in scripts because it
is not immediately obvious what "-c" might do, but I could not find a
long form in the man page.