Add error type, and FFI machinery for it
Noblit will feature a single enum for errors, where the variants may
hold more details if necessary. This has worked well for me in the past,
I see no reason to change it here. The variants may contain &'static str
for details like format errors, or dynamic strings for errors dependent
on user data. At this point I do not have a need for more structured
data than just a string.

Then for the FFI, we can translate those to status codes. The enum type
in C does not have a specified memory layout as far as I can tell, so we
use u32 instead with a bunch of #defines. This is only for documentation
purposes anyway.

We can't return more details than the status code, so that we will store
in the context, with a getter function to get the details in case an
error occurred. The message is exposed as UTF-8 bytes, and the caller
gets to borrow those until the next call.