Run Nix only once on CI to set up environment
The initial call to nix-build on CI takes about 40 seconds. The next one
takes 1 second (just for lib path, should be a no-op), and the one after
that takes 5 seconds. I expect we can save almost 6 seconds by doing
everything in a single nix-build call, and saving the path. It also
allows us to get rid of one file.