Expose query evaluator through FFI
It took me a lot of fiddling to get to this point, but I think I managed
to find a way to do it that is reasonably safe and reasonably free of
boilerplate. The idea is to introduce a struct "Contextual<T>", which
stores a pointer to the context, as well as some value we want to track.
We store the pointer to the context, rather than having the foreign
caller pass it, to ensure that the value and the context always go
together; foreign callers may have multiple contexts in scope. With the
context, we know the type of the database in the context, and from that
we can determine the type of the thing in the "Contextual<T>" box. With
some macros we can then have generic implementation functions, and have
the macro call that in such a way that the types get inferred correctly.