Give cursor a lifetime, add Cursor::from_bytes
This is useful for parsing, where we get a byte array of the input
query, but only for a limited lifetime. By using the cursor, and not
assuming a byte array, we can still stream queries easily (e.g. by
reading them from stdin), but we can also consume byte slices. This is
nothing more than what "io::Cursor" already does, only it makes the
types a bit more tractable.