Go the monomorphization way after all
I wanted to work with trait objects, instead of the generic functions,
because that would simplify the types, and because the type parameters
of an Evaluator are tied to that of the Database, but generics are lost
across FFI boundaries, so I thought that would become a mess, and it
would be better to just have a single type for everything, and use trait
objects.

But it was never going to be *that* simple, because there is the
distinction between "Store" and "StoreMut", and "Heap" and "HeapMut", so
we would still need to dispatch on those. No big deal, but ... then if
you have a "Box<dyn StoreMut>", but you need a "Box<dyn Store>", then we
need to box the box, which feels ridiculous.

So I changed my mind again. Probably all implementation functions should
be generic over the heap and store, and the FFI entry point would
dispatch on the type using an enum, and call the generic function. That
simplifies the implementation functions, and they get to continue to
precisely express the requirements on the store and heap in the types.
To then dispatch and call such a generic function, we need higher-kinded
types which Rust does not have, but that can be worked around by using a
macro instead of a function.