Begin working on build automation script
There are quite a few steps to checking Noblit. It's not just "cargo
build", "cargo test", there are the golden tests, typechecking the
Python code, etc. All of these steps are present in the Travis config,
but I would like to run the steps locally too. So I want some kind of
script that runs these steps sequentially, preferably in the same way
that it works on CI. We can use Nix for this. As a first step, add a
script that sets up the environment such that we only have dependencies
from the Nix devenv and nothing else.