Make build script interpreter for build file
This makes the build file a bit more readable; it's just strings,
instead of Python lists of strings. The build script then doesn't have
to define a step type, it can interpret the build file on the go.

Of course, secretely this is a not-invented-here variant of Bash. Soon
feature creep of variables will slip in, to e.g. avoid repeating the
Rust nightly version. But being custom has its advantages: I can change
it to behave exactly the way I like. For example, to run only some steps
locally with a filter. And it does not have as many features, so it does
not have as many pitfalls.