Revert upgrade of pinned Nixpkgs version
This reverts commit fc430e104bb151bf437ac1b1e871d89238373f53 (again),
because apparently, Rustup is still broken. However, I am not sure that
it does not provide a "cargo" binary; the Nix profile contains one, so
it should be there on CI. I think it may be a misleading message on CI,
where it says

    error: command failed: 'cargo'
    error: caused by: No such file or directory (os error 2)

It may be Cargo printing the no such file or directory message? Anyhow,
I do not want to look into this right now.