Add smarter structure-aware fuzzer for htree
I already had two structure-aware fuzzers for the htree: htree_insert_-
value, and htree_insert_entity. For the structure-aware part, I wrote
a byte-array-to-struct decoder by hand. But the "arbitrary" crate
recently got an overhaul, and it does basically the same thing, but it
does two things better:

 * It uses length suffixes instead of length prefixes. Apparently that
   works better with libFuzzer's mutations, which makes sense to me,
   because if a mutation messes up a length prefix, then everything
   behind gets a new meaning, so putting the length at the end may help.

 * It has #[derive(Arbitrary)], which is really *really* nice to avoid
   boilerplate.

This is not without downsides though: adding a dependency on the
"arbitrary" crate, and its deriving functionality especially, brings in
many new dependencies. Fortunately these are only required for fuzzing,
and not needed for Noblit itself, not even for building it, so I think
this is acceptable.

The new unified "htree_insert" fuzzer supersedes both existing fuzzers;
it now generates both unique entities and unique values in the same
fuzzer. Also, the value fuzzer now generates both integers and byte
strings, not only byte strings. This is mainly because Arbitrary makes
it so easy to do this, and previously I just had not gotten around to
writing it.

In terms of coverage: after 30 minutes or so, the new fuzzer reaches a
higher feature and coverage count than either previous fuzzer, but that
is no surprise, because the new (derived) test case generation now has a
lot more branches in it. So I am not sure what the true effect on
coverage is. I assume both get to 100%.

In terms of speed: runs per second is in the same ballpark as the old
fuzzers for a release build with debug assertions enabled.