Fix type confusion bug in FFI
I changed the impl function from returning a T to returning a Result<T>.
Because the type parameter of the Contextual<T> is inferred, that was
boxing a result, rather than the ok value of the result. And then the
free function assumed a T rather than a Result<T>. The fix is to not box
the result; we should handle the error.

This is tricky though; ideally a type system would prevent such a bug,
but in this case we need to work with macros due to the lack of
higher-kinded typed, at the cost of having to rely a lot on inference,
which can easily go wrong, as demonstrated by this bug.

I did add a few more type hints to make it more difficult to introduce
type confusion bugs, but it does not eliminate all opportunities for
error.