Exercise all indexes in htree_insert fuzzer
Previously it was hard-coded to the EAVT datom ordering. Now we test all
of them. (This does make the previous corpus useless, a disadvantage of
deriving Arbitrary: you can't do migrations.) I don't expect this to
make a huge differce, because the datom ordering is relatively simple,
so the htree was already covered well. Still, it is good to exercise all
code paths, so now we do.