Add contribution guide
At this point Noblit is not open to external contributions. I thought I had
clarified that, but that was only in the Mindec repository, not for Noblit. To
avoid disappointing people who want to contribute, I should clarify ahead of
time that contributions will not be accepted.

Noblit is "source available", under a free software license, and developed in
the open, but it is not open for anybody to contribute at this time.