Delete large-value non-goal
It is still not an explicit goal to store large values, but the current
format happens to support it, so there is no reason to avoid it.
Breaking up large values can be useful for many things, for example for
paging, but due to the append-only value heap, that is less of a
concern. So we can do large values. Noblit is not optimized for it, but
it supports them.