Delete outdated example
This was an experiment to settle on a design for what queries would look
like from Python. But by now, golden tests exists, and the Python client
library is in progress, so a real example that can be tested on CI is
close, whereas the example that I am deleting now was only hypothetical.
Let's remove it, so it does not cause confusion.