Begin working on a fuzz corpus packer
Currently, fuzzing on CI is not terribly useful because the corpus is
not preserved between runs, and the time is quite short. It would be
nice to kick-start the process, or even to save and resume between runs.
I could tar the corpus and pull it through e.g. Brotli, but that would
be wasteful, because the file names are hashes, which are random, so
they compress poorly. But the file names aren't even needed, we can
compute them at extract time. Also, I suspect that sorting the entries
will provide a much better compression ratio, because backreferences
work better.