Pack fuzz input lengths a bit tighter
It gets my full htree_insert corpus down from 75 MB to 0.6 MB now, which
is almost small enough to justify including it in the repository if
necessary. But the corpus still contains many redundant entries, so
after minimizing the corpus, perhaps it will be small enough to be worth
including by default, or maybe to host it somewhere.