Begin working on haveibeenpwned example
Inspired by the this blog post that I read recently [1], I think that
testing whether a password is in the Have I Been Pwned database would be
a good simple example for Noblit that can serve as a stress test at the
same time. The post suggests that a B-tree is a good datastructure for
doing a set membership test quickly on a large set. Noblit has something
like B-trees. So simply putting all of this in a Noblit database should
be a reasonably quick way of testing whether the password is present,
and it is a good test for databases that do not fit in RAM.*

[1]: http://stryku.pl/poetry/okon.php

* The uncompressed text file is 24.4 GB, which is more than the amount
  of RAM in my laptop. But it does store the hashes as hex, and counts
  as text, which is wasteful -- simply storing that as binary will cut
  the size in half, and then it *would* fit in my RAM. But Noblit has
  96 bytes overhead per datom, excluding the tree overhead even, which
  is more than the 40 bytes of hexadecimal SHA1, so the database is not
  going to fit in RAM on my laptop.