Parse passwords file, insert into db in batches
I still don't have a disk-backed database, so this is only inserting
until the store exceeds 1G, to ensure my machine does not go out of
memory. This limit can be lifted later, for now I just want to set up
the program to ingest the data, so I can query the database.