Implement password-present query
If I now test for "hello" (aaf4c61ddcc5e8a2dabede0f3b482cd9aea9434d) on
the truncated test database, it finds that it is present. If I test for
"flubblewub" (db1610b32c3328942be8dfe4fbf23f0b86243263), it is not
present. So it looks like it works, yay!

Next up: a file-backed heap and store, to work with data that does not
fit in RAM.