Finish support for reading from mmapped database
I changed the endianness of all integers in the heap earlier, so now the
haveibeenpwned database that I still have sitting on my disk can't be
read any more and hits an assertion. But it does hit that assertion a
lot earlier with the mmap version than with the full-read version now.