Optimize bounds checks in slice_8
These custom panic messages are nasty, it means that even though the
code would crash if any of the indices was out of bounds, the compiler
can't optimize away the checks, because that would change the crash
message. I stumbled upon a bunch of cmps in heap::get_u64 while
profiling, while actually I want get_u64 to compile to a normal load.
The code is written to avoid a transmute, but it is a convoluted way of
saying that I want an unaligned load. And actually it could be an
aligned load; the format guarantees that, only the heap must be loaded
at the right offset then.