Make optimizer iterations adaptive
Without this, it can take hours to optimize a query whose running time
is in the millisecond range; this makes it a bit faster.

I also had this idea: being precise in the measurement is only important
near the leaves of the tree, and when we get close to those, in the
beginning we can be more sloppy, because if the difference between paths
is small, we will try all of them anyway if the next step adds more time
than the difference between two nodes at the same depth. So the number
of iterations could also adapt to the depth in the tree. I did not
implement that yet though, let's see how far we get with this simple
thing.