Begin working on some async key-value store ideas
The hitchhiker tree is great. But coupling it so tightly to datoms ...
it may be great for other things too, maybe i should have one for just
key-value purposes. (I want to build a file system on top of this at
some point.) And while we're at it, all IO should be async through
queues, so it maps to NVME and the way SSDs work, but surprisingly, also
remove stores and caches and stuff should fit into this protocol neatly.