Toy around with async
After thinking about this for quite some time (months!), probably the
best way for Noblit to leverage io_uring, is Rust's async after all. It
is complex and I would have liked to stay away from it, but async _is_
complex, and making something high-performance means it needs to be more
complex than sequential code. (For example, we have to gather all reads,
kick them off as a batch, and then continue. It is not enough for IO to
merely be non-blocking, we want to use parallelism of the underlying
device.)