Pin Nixpkgs to latest 23.05
I'm doing this across most of my repositories, so I can have a single
copy of Nixpkgs in my store, since all repositories pin to the same
version.

The fuzzing part of the build is broken, but not sure if that is caused
by this update or if it was already broken.