Think about a Hitchhiker-tree key-value store again
I want to build an encrypted secret store. I need some kind of key-value
store to back it. I want it to be distributed with a nice way to sync
between nodes, and sending just all new pages is a very nice way of
doing that. Let's explore doing it based on Noblit.