Add new ideas: a copying generational GC store
I wanted to start implementing today, but I was not satisfied with the
way of storing values. I did some brainstorming session with Claude, and
after several side tracks, this design seems simple enough that it's not
too difficult to implement (and implement correctly!), while it also
doesn't do anythign that looks obviously silly from a performance or
storage space point of view. Even the write amplification is I think
pretty good.